Extract InfoWindow components to separate file
- Move InfoWindow, InfoWindowFactory, and related interfaces to infowindow.tsx
- Extract createHeaderContent function for better organization
- Update roadmap.tsx to import InfoWindow components
- Simplify roadmap.tsx by removing InfoWindow-related code
